+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Add "warn" issue level and update issue messages
 
+- Change ObjectSchema extend method to accept ObjectSchema instance or shape
+
 ---
 
 ## v0.4.0 – 2025-05-18