+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## v0.5.0 – 2025-05-19
 
 - Introduce fallback method and FallbackSchema